+++ CHANGELOG.markdown
@@ -23,6 +23,7 @@
 * Added `Strict(strict)` for ad hoc overloading of conversions between strict and lazy variants of `ByteString` and `Text`.
 * Bug fixes for `tugTo` and `jerkTo`.
 * These no longer traverse in the wrong direction: `scanl1Of`, `scanr1Of`, `mapAccumLOf`, and `mapAccumROf`
+* Added `embedding` to `Control.Lens.Projection`.
 
 [3.6](https://github.com/ekmett/lens/issues?milestone=9&state=closed)
 ---